[ISSUE-590] Convert BranchUtils to Java
